enable gentle fair sleepers switch thanks to Neobuddy
Sleeper Fairness is a concept used by CFS which treat sleeping/waiting tasks as if they were
in a run queue. This implies tasks which spend most of the time waiting for an user input
and such will get a fair share of CPU when they need it. Disabling Gentle Fair Sleepers
could improve UI responsiveness.

Signed-off-by: mostafa-z <mostafazarghami@gmail.com>