makefile: enable some optimization flag. read more
-funroll-loops

    Unroll loops whose number of iterations can be determined at compile time or upon entry
to the loop. -funroll-loops implies -frerun-cse-after-loop, -fweb and -frename-registers. It
also turns on complete loop peeling (i.e. complete removal of loops with a small constant
number of iterations). This option makes code larger, and may or may not make it run faster.

-fsingle-precision-constant

    Treat floating-point constants as single precision instead of implicitly converting them
to double-precision constants.

-fgcse-lm

    When -fgcse-lm is enabled, global common subexpression elimination attempts to move
loads that are only killed by stores into themselves. This allows a loop containing a
load/store sequence to be changed to a load outside the loop, and a copy/store within the
loop.

-fgcse-sm

    When -fgcse-sm is enabled, a store motion pass is run after global common subexpression
elimination. This pass attempts to move stores out of loops. When used in conjunction with
-fgcse-lm, loops containing a load/store sequence can be changed to a load before the loop
and a store after the loop.

-ftree-vectorize

    Perform vectorization on trees. This flag enables -ftree-loop-vectorize and
-ftree-slp-vectorize if not explicitly specified.

-floop-nest-optimize

    Enable the isl based loop nest optimizer. This is a generic loop nest optimizer based on
the Pluto optimization algorithms. It calculates a loop structure optimized for
data-locality and parallelism. This option is experimental.

-fpredictive-commoning

    Perform predictive commoning optimization, i.e., reusing computations (especially memory
loads and stores) performed in previous iterations of loops.

Signed-off-by: mostafa-z <mostafazarghami@gmail.com>