cpufreq: gabriel*: Skip timer when policy->min == policy->max
When the minfreq is set to the maxfreq, there is nothing to really do: the
CPU is already running at its maxfreq. In this scenario, just do nothing
and re-arm the timer.

This scenario occurs when the screen is turned on due to the cpu_input_boost
driver (it sets policy->min to policy->max for a few seconds).

original-commit
https://github.com/nikhil18/Lightning_kernel_oneplus5/commit/62924b2054e87ee8865f0dcaf4783fb9c73fcc82

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Joe Maples <joe@frap129.org>
Signed-off-by: mostafa-z <mostafazarghami@gmail.com>